review register page

+++ src/Controller/UserHomeController.php
@@ -15,39 +15,40 @@ class UserHomeController extends AbstractController
     #[Route('/user/home', name: 'app_user_home')]
     public function index(ReservationsRepository $reservationsRepository, ManagerRegistry $getDoctrine): Response
     {
+        /* INIT VARIABLES */
+        $review = new ReviewsType();
         $dateReservation = null;
+        $todayDate = date('d/m/y');
         $display = false;
         $email = $this->getUser()->getUserIdentifier();
         $reservations = $reservationsRepository->findBy([
             'email' => $email,
         ]);
-        $todayDate = date('d/m/y');
-        if (isset($reservations[0])) {
-            $dateReservation = $reservations[0]->getDateofreservation();
+
+        /* On vérifie si user à une réservation passée */
+        if($reservations !== []) {
+            if (isset($reservations[0])) {
+                $dateReservation = $reservations[0]->getDateofreservation();
+                $newDateReservation = strtotime($dateReservation);
+                if ($newDateReservation < time()) {
+                    $display = true;
+                    if(isset($_COOKIE['review']) && ($_COOKIE['review'] !== null) && ($_COOKIE['review'] !== '')) {
+                        $review->setDate($todayDate);
+                        $review->setEmail($email);
+                        $review->setReview($_COOKIE['review']);
+                
+                        $entityManager = $getDoctrine->getManager();
+                        $entityManager->persist($review);
+                        $entityManager->remove($reservations[0]);
+                        $entityManager->flush();
+            
+                        unset($_COOKIE['review']);
+                        setcookie('review', '', time() - 3600, '/');
+                    }
+                }
+            }
         }
-
-        $newDateReservation = strtotime($dateReservation);
-        if ($newDateReservation < time()) {
-            $display = true;
-        }
-        // /* REVIEW COOKIE */
-        $review = new ReviewsType();
-        if(isset($_COOKIE['review']) && ($_COOKIE['review'] !== null) && ($_COOKIE['review'] !== '')) {
-            $review->setDate($todayDate);
-            $review->setEmail($email);
-            $review->setReview($_COOKIE['review']);
-    
-            $entityManager = $getDoctrine->getManager();
-            $entityManager->persist($review);
-            $entityManager->remove($reservations[0]);
-            $entityManager->flush();
-
-            unset($_COOKIE['review']);
-            setcookie('review', '', time() - 3600, '/');
-
-
-        }
-
+        
         return $this->render('user_home/index.html.twig', [
             'display' => $display,
         ]);
